requirements.txt
streamlit==1.36.0
scikit-learn==1.3.2
numpy==1.24.4
matplotlib==3.7.3
pillow==10.0.1
opencv-python-headless==4.8.1.78
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.3
